requirements.txt
pandas==1.1.5
streamlit==1.14.0
numpy==1.19.5
plotly==4.14.3
protobuf~=3.19.0
Pillow
typing_extensions
